requirements.txt
djangorestframework
Pillow
django-filter